Merge the rlp_serialization module in eth_types
Having 'rlp_serialization' and 'eth_types' as separate modules is
too dangerous. It's quite easy for someone to include only the types
module and to invoke the default RLP serialization that doesn't do
the right thing for MDigest and Stint values.

Perhaps, we should remove the default serialization altogether and
leave it only for tuples, but for now the modified module layout here
provides better safety.